Stop using inheritance and tying myself in knots, do proper pipeline creation params for video:: objects, the video::IPipeline didn't provide anything useful!
